rework home page
