--- client/components/Home/AnonHome.tsx
@@ -1,25 +1,0 @@
-'use client';
-
-import Hero from "@components/Home/Hero";
-
-/**
- * Header:
- * - Tracking grades alone is out, tracking goals is in.
- * - Why only reflect on the past when school is about your future?
- * - School is a marathon, not a walk-and-then-sprint-the-last-200-meters-a-thon.
- * - Last minute exam anxiety is so last year.
- * - Know what you need to get on the final -- and everything else -- from minute one.
- *
- * Body:
- * - Instead of relying on grades, math, and chaotic spreadsheets to track your academic goals, let us do the math. CourseFull lets you focus on your coursework while knowing just how much breathing room you have, or by how much you need to catch up.
- * - CourseFull tracks your grades, and what you need to achieve your goals, all semester long.
- */
-
-export default function AnonHome() {
-
-    return (
-        <div>
-            <Hero />
-        </div>
-    );
-}
